Change name of checkoperation Document and name the parsers Fix a bug in type casting in expressions

+++ src/variables/EVInt.cs
@@ -29,7 +29,7 @@ namespace E_Lang.variables
 
     public override dynamic Get()
     {
-      return (decimal) value;
+      return value;
     }
 
     public override EVariable Set(dynamic setto)
